// types.ts
export type AppID = 'about' | 'projects' | 'contact' | 'settings' | 'ai_assistant' | 'terminal' | 'reader' | 'file_explorer' | 'hire_me' | 'tic_tac_toe' | 'games_folder' | 'rock_paper_scissors' | 'connect_four' | 'guess_the_number' | 'hangman' | 'skills_playground' | 'work_with_me' | 'secret_projects' | 'business_card' | 'skills_interests' | 'articles' | 'editor' | 'data_wave_dashboard' | 'case_studies' | 'astro_viewer' | 'astro_tracker';
// ...
export interface Project {
  category: string;
  title: string;
  description: string;
  features?: string[];
  contributions?: string[];
  tasks?: string[];
  links: {
    live?: string;
    github?: string;
    website?: string;
    instagram?: string;
  };
  icon: string;
  appId?: AppID;
}

export interface CaseStudy {
  id: string;
  title: string;
  subtitle: string;
  date: string;
  project: string;
  content: string;
  snippet: string;
}
// ...
export interface VFSNode {
    type: 'file' | 'directory';
    name: string;
}
export interface VFSFile extends VFSNode {
    type: 'file';
    content: string;
    isLink?: boolean;
    isShortcut?: boolean;
    appId?: AppID;
}
export interface VFSDirectory extends VFSNode {
    type: 'directory';
    children: { [key: string]: VFSNode };
}